+++ packages/opencensus-instrumentation-grpc/test/test-grpc.ts
@@ -43,6 +43,7 @@ class TestExporter implements StatsEventListener {
   }
 
   start(): void {}
+  stop(): void {}
 
   clean() {
     this.registeredViews = [];