+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - Enforce `--strictNullChecks` and `--noUnusedLocals` Compiler Options on [opencensus-instrumentation-http], [opencensus-instrumentation-grpc] and [opencensus-propagation-tracecontext] packages.
 - Enforce `--strictNullChecks` and `--noUnusedLocals` Compiler Options on  [opencensus-exporter-zipkin] packages.
 - Add NoRecordRootSpan, NoRecordSpan and NoRecordSpanBase.
+- Add an API `globalStats.unregisterExporter()`.
 
 ## 0.0.9 - 2019-02-12
 - Add Metrics API.